Try to optimize network_transmission, assuming the problem was CPU usage
The download speed seemed reasonable to me, but this does improve it slightly
(from 1.2MB/s to 2MB/s, testing connected via my phone). The optimization here
is assuming that redrawing the window is causing another event to be pumped,
and so making this loop a busy loop that is pumping itself.