Update a game's ID when advancing to the next scenario.
On receiving "store_next_scenario", increment wesnothd's game id counter and assign the game the new id.
Currently this is relevant for when an MP campaign advances to the next scenario, in order to avoid primary key conflicts as well as insert a new game_info row for the new scenario.

Also fixes #4281 by adding a db call in "store_next_scenario" to insert the next scenario's game_info row.