UtBS: Replacing ^Uf mushrooms with ^Tf mushrooms
The code for flooding in 08_Out_of_the_Frying_Pan uses a list of impassible terrains,
anything not in that list can flood so there's no change needed to that code.

The Sun Sylph's animation code now recognises both types of mushroom, in case the
Quenoth elves are used in UMC campaigns.