UtBS S09: Non-merfolk get in the ships before they sail off
At the end of S09, the heroes have four ships. The narrative is that all the
elves are on the ships, but the in-game visuals were that the ships move while
all the player's units are standing on the shore.

The heal=yes isn't needed since 74c2398f6, but adding it here avoids
a possible bug if this commit is backported to 1.14.