Help for the ^Uf mushrooms terrains, including an editor-only hint
As the terrain is hidden in help, these hints can only be seen by
using the "Terrain Description" command on a ^Uf or ^Ufi mushroom
hex, and then switching to the overlay's own page.

A hint for players about how these mushrooms behave, and an editor-only
hint about replacing the ^Uf mushroom terrains with ^Tf.