TRoW S04a S04b: From Midlands, make Swamp the next scenario
The Midlands is the more challenging option, but means missing out on
an auto-recalled loyal white mage. The journey map also shows that The
Oldwood is east of The Swamp of Esten, and it makes little sense that
they reach Oldwood without passing through the swamp. Solve both
problems by making The Midlands an optional detour which still ends up
going through the swamp.

If the player took the extra XP by going through The Midlands, the
turn limit on The Swamp is much tighter. Sir Ruddry will appear during
the start event, tell Haldric about the road ahead, and then die - Sir
Ladoc gets an exposition that only one of the two knights can survive.

I've left the scenario number unchanged, although this means that 04b
will occur (if the player chooses to go there) before 04a.