fs: Drop filesystem::init() and Windows XP-specific code in it
The function's only purpose so far is to initialize a function pointer
to a kernel32.dll export (GetFinalPathNameByHandleW) that isn't present
on versions before Windows Vista. However, this development branch
explicitly bumps the minimum build target to Windows 7, rendering the
code redundant.

This also gets rid of a rather bothersome GCC warning involving the
FARPROC pointer cast used to store the result of GetProcAddress().