Tutorial S02: Show "recruit the right units" when the player has gold
It's confusing when it pops up on turn 12, regardless of the game state. My
leader is normally on the other side of the river from my keep, and I always
wonder "what about my play triggered that?"

Move the message to be part of the "I now have enough gold to recruit more
units" hint.