TSG: remove not anymore needed macro
The alternative message was needed for the scenario where Deoran is not on the map.
This has been removed, and in the other which is left (S8b), both units won't be around.

The only places left for this message is the last level in path B and the lasts ind path A
It's not necessary that there is an alternative message, this is only used there because
it exists already – both would fit.

[ci skip]