[print]: Center the message within the map area
In master, this just moves the message down by the height of the top bar.

In themes where the left end of the map isn't at x=0, this makes [print]
centered within the main map, rather than centered within the rectangle
with corners (0,0) and (bottom right corner of the main map), regardless
of where the top and left ends of the map are with respect to the screen.

Currently, the [print] messages are vertically centered, whereas the
"Planning mode activated" message is vertically at one-third of the
height. We should perhaps make these consistent, but this commit
doesn't change that.