Editor help: remove whitespace from ends of strings
Note: this does not include a fix for #4484, that a double-quote in a _<< >>
delimited string causes a broken msgid. Wondering whether to remove that from
this text, but would prefer to add the feature to wmlxgettext.

Follows the comments in #4480, about how to handle whitespace at the beginning
and end of translatable strings.

This is text that was added after the last pot-update, so updating it
again now will hopefully be before the translators start on it. Some
text that was split into paragraphs is now combined in to one block,
other parts are still split but with non-translatable whitespace; the splits
will hopefully separate the scenario editor parts which are going to change
again before 1.16.0.

The wmlindent override is because the logic for multiline strings isn't working
for << >> delimited strings.