Translations: drop languages with 3 or less translated strings
* Old English (ang)
* Friulian (fur_IT)
* Asturian (ast)

The "Old English (Latin chars) (ang@latin)" translation is not dropped,
which is why Old English is still listed in about_i18n.cfg.

The manpage uses ang_EN in the example of the command-line --language
option, which is then copied in to other translations. That isn't changed
in this commit.

Although the po4a scripts are updated, the update-po4a target currently
fails because of a typo in the pt_BR translation.

Two scripts have a special case for ignoring ang_GB, both of these are
left unchanged on the assumption that old versions of Wesnoth may be
parsed with the latest version of the tools:
* data/tools/unit_tree/helpers.py
* data/tools/wmlunits