Display left-facing Delfador portraits on the right.
Some of Delfador's portraits face to the left. A cosmetic change only, but we prefer to such portraits to be displayed to the right of the dialogue so that they face the 'middle'. It also gives impression of a conversation with a right-facing portrait displayed (by default) on the left.

Resolves #4345.

[ci skip]