fs: Consider platform differences when enumerating alternate save paths
The original code by octalot only works correctly on X11 when using the
recommended fdo layout. Windows and macOS need their own logic to build
candidate paths.

The code is still very optimistic about this whole thing -- see the
comment. Ideally we need to refactor the system to determine user data
path defaults so we can do it as many times as necessary even if we've
already determined our current user data path, but that seems like a lot
of work for a worst-case scenario that's not gonna crop up very often.

Also, I have no idea if the macOS alternative works, or what will happen
on iOS for that matter. It should at least compile.