Disambiguate uses of the Play button label
This extends to a few other button labels used for replay control. While
checking if commit 46dbbc06c967bfc8ee5acb6525e1f1bd96b53005 was fit for
backporting to 1.14 I found out that "Play" was already in use
exclusively for one of the replay control buttons (which means, no, it
can't be backported). This makes the disambiguation markers absolutely
necessary.

To give a more practical example of why this is a big deal, in Spanish,
"Play" would be translated as "Jugar" in the context of the Campaigns
menu, and "Reproducir" in the context of a replay (or movie). The
Spanish translation in fact already uses the latter in both 1.14 and
master.

[ci skip]