Accept heal_full in [effect]apply_to=variation
I thought this had been done in 217992a843f1ca9fc62eda02d0e376619ab786f7 but it turns out that was in the rejected version of the commit and didn't find its way into the final version.
Instead of reverting the schema and changelog mentions of it, I've just added the feature.