gui2/campaign_selection: Consider untranslated search terms
As I said in 739e2608db6b312a8e78b830e820a7970b52782c, the abbreviations
and names of campaigns in English tend to be better known by the
community, so they are also useful to consider when searching through
the list.

The description's untranslated version is also included in this even
though descriptions are longer and a bit more expensive to sift through.
It may still come in handy in cases where proper names get inflected or
transliterated I guess.