gui2: Delegate mouse cursor reset to modal dialogs
This is a backport to 1.14 of two commits from master:

 * gui2: Delegate mouse cursor reset to modal dialogs
   47ddf8877f5bb481b933b2a3dbc012d4c156d69a
 * gui2: Fix cursor setter going out of scope as soon as it's created
   751043a692920a2e95a79a3a7c488eeda2b16ede

Reason being that in the original, the first commit had a major bug
(issue #4618).

Currently it's the gui2::window class' responsibility, by inheriting
from cursor::setter (yes) and offering absolutely no way to control it.
Unfortunately, it turns out that this can cause issues with tooltips
resetting the cursor at unexpected times when they get
displayed/hidden while having widgets change the cursor depending on
whether they have the mouse focus or not.

Right now the only modeless dialog we have is the tooltip dialog, and
really there's no reason to say that modeless dialogs should have the
ability to reset the cursor without explicitly asking for it. As for
modal dialogs... I don't entirely agree that they should do it either,
but for tradition's sake, let's just keep the current behaviour with
them.

(In an ideal world the cursor business would be managed by the window
managaer IMO, but I'm not ready to argue my case.)