gui2/text_box: Implement I-beam cursor
This is a backport to 1.14 of multiple commits from master:

 * gui2/text_box: Implement I-beam cursor
   d7f2595393e2744d87ee504557b92264fca29741
 * Don't forcefully disable colour cursors
   fc04268aa4eb725433d1db458973f1d187f73598
 * Set the I-beam cursor's hotspot properly
   c31047af4723c76dcfb3561d77a3d6c211c2994c

This cursor is active when the widget has the mouse focus. You know,
like textboxes are wont to do anywhere else. Took long enough, although
there was an interaction issue with tooltips fixed in the previous
commit.

I still need to figure out how to make this work with GUI1 textboxes
(e.g. in-game console).

The colour version of the cursor was kindly provided by LordBob, and the
B&W version is my own.