Add trackviewer, remove trackplacer
Trackviewer shows a journey without needing to start Wesnoth and refresh the
cache. Combined with tmx_trackplacer, the python3 versions now have
feature-parity with the old trackplacer, without the reliance on python2 or
pygtk (fixes #4365).

At least on Linux you can have both Tiled and this open on the same file. Save
the file in Tiled, alt+tab to this, and press enter to reload the file.