Separate unit test statuses for WML exceptions and strict warnings (fixes #4471)
Backported to 1.14 so that this mechanism can be used instead of a793bce96's
method of running the cve_2018_1999023 test in non-strict mode, which means
that the test now passes without making equivalent-to-a793bce96 changes to
projectfiles/VC12/WML_tests.cmd and projectfiles/VC16/WML_tests.cmd.

This was part of working out whether a subset of the "fail" tests could be run
in one Wesnoth instance. To run a test that returns TEST_FAIL_BROKE_STRICT with
any other test would require a mechanism to reset lg::broke_strict()'s flag.

All tests that fail with an {ASSERT} will also set the lg::broke_strict() flag,
the tests with the new status are only those that would pass without the strict
flag.

In the SUF tests, change a test from fail-on-success to breaks-strict, rename
the formula tests and add some comments. The rename is because "fai" is
"Formula AI", an obsolete name for WFL.

(cherry-picked from commit 7359fddecb8dafdb8bec33660259ec5e2fec4141)