ai unit advancement improvements
1) use ai unit advancement asepect also
   during [event]s fixes #4605

previously advancements of ai units
invoked by [event]s for example by
unstoring a unit with enough xp would
always do a purely random advancement even if
it happend during the ais own turn. Now the
advancemt_aspect is considered. I don't really
know how the code inside ai/ works so i just
took the easiert way to get a advancements_aspect
from the manager. In particular this leaves a
TODO comment about whether the advancemnts_aspect
of advance_unit_at (and thus of
attack_and_advance) can not be removed or not
because i don't know whether the advancement
aspect is effectively a singleton (per side)
or not.

2) support amlas in ai units_advancement_aspect