Remove tests of move_unit's count and move-unit-to-its-own-hex functionality
The count is undocumented on the wiki, and it's causing the schema validation
to fail. The functionality makes sense and I hope it continues to be supported,
but for now let's have a minimal unbreak-the-build change.

Trying to move a unit to the hex occupied by itself causes it to move to a vacant
adjacent hex instead; splitting the move into two separate [move_unit] calls
passes as expected. That sounds reasonable implementation-defined behavior, but
will make a PR to discuss it.