ANL: ai tweaks – don't retreat
In this scenario, the AI can't heal his units, as they destroy buildings
(except for leaders). Thus no sense in keeping units alive.

Disable caste_switch, in case players find a way to abuse it (as they can build castles)

For the Spiders, disable the poison CA, as it's only a monster encounter,
no need to be tactically smart in the long run.