fix another bug in unit sidebar
I detected it by testing weapon-> under_leadership before returning the old function but [damage] and [attacks] are also concerned, if a unit under leadership is selected for movement, if the destination pointed by the cursor is another hex adjacent to the leader, then there will be a different damage value than that obtained when the unit is moved there. This is due to the fact that abil_list is never zero but has self_loc_, and that in the if (self) loop, append is used. I solved the problem using abil_list = list_leadeship ..
to clear the default value. I don't understand why abil_list (self_loc_) could generate this bug, but it should only be used for the declaration of the list and be replaced afterwards.