Simplify help's logic for terrain_types that should be hidden in help
Even when a terrain has hide_help=yes, its help should still be shown if the
player right-clicks on a hex of that type and selects "Terrain description".

This commit changes lines that have inconsistent logic, and the main reason for
doing it was to work out whether that inconsistency was something that needed
to be kept when refactoring the terrain code.

The inconsistency was a bug, help would fail to open the topic.  However, the
bug is so obscure that I only found it when considering the edge case that the
commit is removing, and there's no terrain in core that would trigger it.  The
edge case requires a terrain to be:

* defined by a [terrain_type] tag, so is_combined()==false
* hide_help=yes
* hidden=no (or is the default value)

To replicate the bug, try changing the definition of Gg to have hide_help=yes.