log/windows: Ignore failure to attach or allocate a console
Wine as of version 4.21 just refuses and goes ERROR_ACCESS_DENIED when
trying to allocate a console for a GUI subsystem application. Why? Who
knows. It does have the functionality to run console applications either
using ncurses or its own console implementation, but it won't. It's best
not to dwell on it and just ignore what's supposed to be the unlikeliest
worst case scenario.