EI S11: Trigger the guards if an undisguised unit moves into the throne room
Add an enter_hex event for moving into the throne room, and modify the "attack"
event so that the undisguised units can trigger it too. For reviewing, note that
the diff shows these as if the existing "attack" was changed to a "enter_hex",
but in fact it's the existing `[filter_wml]` moving from "attack" to "enter_hex".

Add L2 Trolls to various lists, as they're used on easy/normal difficulties.

The main bug fixed here is "When a guard sees one of the two units not
disguised as an orc, it chases after the unit, but no other enemies are
alerted. This allows a player to pick off each enemy unit individually, then
release the prisoners."

(cherry picked from commit c66bd63911cda57acd64c9142623c6eda8ab93b5)