Move the macOS builds to the front of the Travis job queue.
For some reason the xcode job easily runs the longest despite being a partial recompile.  Having it also be the final job to run therefore means that the entire build will be waiting for it to complete, whereas having it earlier allows other shorter jobs to be run alongside it.